+++ src/functions/organizations.ts
@@ -1,8 +1,8 @@
-import { BasePageFetchOptions, Organization, User } from "../typings/interfaces";
+import { Article, BasePageFetchOptions, Listing, Organization, User } from "../typings/interfaces";
+import { BaseFetchPageOptions, CategorizedListingOptions, Page } from "../typings/types";
 import DEVUtil from "../structures/DEVUtil";
-import { BaseFetchPageOptions, Page } from "../typings/types";
 
-const { request } = DEVUtil;
+const { request, parseParameters } = DEVUtil;
 
 /**
  * Fetches an organization by it's username
@@ -22,7 +22,33 @@ export async function getOrganizationByUsername(
  * @returns {Promise<Page<User>>}
  */
 export async function getOrganizationsUsers(
-  username:string, options?:BaseFetchPageOptions
+  username: string, options?:BaseFetchPageOptions
 ):Promise<Page<User>> {
   return await request(`/organizations/${username}/users`);
+}
+
+/**
+ * Fetches the listings of an organization by organization username.
+ * @param {string} username 
+ * @param {CategorizedListingOptions} options 
+ * @returns {Promise<Page<Listing & { organization:Organization }>>}
+ */
+export async function getOrganizationsListings(
+  username: string, options?:CategorizedListingOptions
+):Promise<Page<Listing & { organization:Organization }>> {
+  const query = parseParameters(options);
+  return await request(`/organizations/${username}/listings${query}`);
+}
+
+/**
+ * Fetches the articles of an organization by organization username.
+ * @param {string} username 
+ * @param {BaseFetchPageOptions} options 
+ * @returns {Promise<Page<Article>>}
+ */
+export async function getOrganizationsArticles(
+  username: string, options?:BaseFetchPageOptions
+):Promise<Page<Article>> {
+  const query = parseParameters(options);
+  return await request(`/organizations/${username}/articles${query}`);
 }